requirements.txt
g2p-en
inflect
librosa
matplotlib
numba
numpy
pypinyin
pyworld
PyYAML
scikit-learn
scipy
soundfile
tensorboard
tgt
tqdm
unidecode
torch==2.8.0 --index-url https://download.pytorch.org/whl/cu128
grpcio
grpcio-tools
protobuf